pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "geomie3d"
version = "0.0.8"
authors = [
  { name="Kian Wee CHEN", email="chenkianwee@gmail.com" },
]
description = 'Python library for processing 3d geometry'
readme = "README.md"
requires-python = ">=3.10"
classifiers = ["License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
               "Programming Language :: Python :: 3.10",
               "Operating System :: OS Independent"]
dependencies = ['numpy>=1.26.4']

[project.optional-dependencies]
viewer3d = ['pyqtgraph==0.13.7', 
            'PyOpenGL==3.1.7',
            'PyQt6==6.7.1',
            'python-dateutil==2.9.0']

[project.urls]
"Homepage" = "https://github.com/chenkianwee/geomie3d"
"Bug Tracker" = "https://github.com/chenkianwee/geomie3d/issues"